# src/main/resources/openapi.yaml
openapi: 3.0.3
...
info:
  title: Jaqpot API
  version: 1.0.0
  description: 'A modern RESTful API for model management and prediction services, built using Spring Boot and Kotlin. Supports seamless integration with machine learning workflows.'
  contact:
    name: Jaqpot
    url: 'https://jaqpot.org'
    email: upci.ntua@gmail.com
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
security:
  - bearerAuth: [ ]
servers:
  - url: 'https://api.jaqpot.org'
    description: Jaqpot API
paths:
...
  /v1/models:
    post:
      summary: Create a new model
      security:
        - bearerAuth: [ ]
      tags:
        - model
      operationId: createModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Model'
      responses:
        '201':
          description: Model created successfully
        '400':
          description: Invalid input
      x-stoplight:
        id: 9ffjy7o77jc41
  /v1/user/models:
    get:
      x-spring-paginated: true
      summary: Get paginated models
      security:
        - bearerAuth: [ ]
      tags:
        - model
      operationId: getModels
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
            example: [ "field1|asc", "field2|desc" ]
      responses:
        '200':
          description: Paginated list of models
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ModelSummary'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  pageSize:
                    type: integer
                  pageNumber:
                    type: integer
        '400':
          description: Invalid input
  /v1/models/search:
    get:
      x-spring-paginated: true
      summary: Search for models
      security:
        - bearerAuth: [ ]
      tags:
        - model
      operationId: searchModels
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Paginated list of models
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ModelSummary'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  pageSize:
                    type: integer
                  pageNumber:
                    type: integer
        '400':
          description: Invalid input
  /v1/user/shared-models:
    get:
      x-spring-paginated: true
      summary: Get paginated shared models
      security:
        - bearerAuth: [ ]
      tags:
        - model
      operationId: getSharedModels
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: array
            items:
              type: string
            example: [ "field1|asc", "field2|desc" ]
        - name: organizationId
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Paginated list of shared models
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ModelSummary'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  pageSize:
                    type: integer
                  pageNumber:
                    type: integer
        '400':
          description: Invalid input
  '/v1/models/{id}':
    get:
      summary: Get a Model
      description: Retrieve a single model by its ID
      security:
        - bearerAuth: [ ]
      tags:
        - model
      operationId: getModelById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the model to retrieve
          schema:
            type: integer
            format: int64
            example: 0
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '404':
          description: Model not found
    delete:
      summary: Delete a Model
      description: Delete a single model by its ID
      security:
        - bearerAuth: [ ]
      tags:
        - model
      operationId: deleteModelById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the model to delete
          schema:
            type: integer
            format: int64
            example: 0
      responses:
        '204':
          description: Model deleted successfully
        '404':
          description: Model not found
  '/v1/models/legacy/{id}':
    get:
      summary: Get a legacy model
      description: Retrieve a single model by its ID
      security:
        - bearerAuth: [ ]
      tags:
        - model
      operationId: getLegacyModelById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the model to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '404':
          description: Model not found
  /v1/models/{modelId}/predict:
    post:
      summary: Predict with Model
      description: Submit a dataset for prediction using a specific model
      security:
        - bearerAuth: [ ]
      tags:
        - model
      operationId: predictWithModel
      parameters:
        - name: modelId
          in: path
          required: true
          description: The ID of the model to use for prediction
          schema:
            type: integer
            format: int64
            example: 0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dataset'
      responses:
        '201':
          description: Prediction created successfully
        '400':
          description: Invalid Request
        '404':
          description: Model not found
        '500':
          description: Internal Server Error
  /v1/models/{modelId}/predict/csv:
    post:
      summary: Predict using CSV with Model
      description: Submit a dataset for prediction using a specific model
      security:
        - bearerAuth: [ ]
      tags:
        - model
      operationId: predictWithModelCSV
      parameters:
        - name: modelId
          in: path
          required: true
          description: The ID of the model to use for prediction
          schema:
            type: integer
            format: int64
            example: 0
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasetCSV'
      responses:
        '201':
          description: Prediction created successfully
        '400':
          description: Invalid Request
        '404':
          description: Model not found
        '500':
          description: Internal Server Error
  /v1/models/{id}/partial:
    patch:
      summary: Partially update specific fields of a model
      operationId: partiallyUpdateModel
      tags:
        - model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - visibility
                - task
              properties:
                name:
                  type: string
                  minLength: 3
                  maxLength: 255
                description:
                  type: string
                  minLength: 3
                  maxLength: 50000
                visibility:
                  $ref: '#/components/schemas/ModelVisibility'
                task:
                  $ref: '#/components/schemas/ModelTask'
                tags:
                  type: string
                  maxLength: 1000
                sharedWithOrganizationIds:
                  type: array
                  items:
                    type: integer
                    format: int64
      responses:
        '200':
          description: Model fields updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
        '404':
          description: Model not found
        '400':
          description: Invalid input
...
  '/v1/datasets/{id}':
    get:
      summary: Get a Dataset
      description: Retrieve a single dataset by its ID
      security:
        - bearerAuth: [ ]
      tags:
        - dataset
      operationId: getDatasetById
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the dataset to retrieve
          schema:
            type: integer
            format: int64
            example: 0
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
        '404':
          description: Model not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Model:
      type: object
      required:
        - jaqpotpyVersion
        - libraries
        - dependentFeatures
        - independentFeatures
        - rawModel
        - name
        - visibility
        - type
        - task
      properties:
        id:
          type: integer
          format: int64
          example: 0
        name:
          type: string
          minLength: 3
          maxLength: 255
          example: My Model
        description:
          type: string
          minLength: 3
          maxLength: 50000
          example: A description of your model
        type:
          $ref: '#/components/schemas/ModelType'
        jaqpotpyVersion:
          type: string
          example: 1.0.0
        doas:
          type: array
          items:
            $ref: '#/components/schemas/Doa'
          x-field-extra-annotation: "@get:org.jaqpot.api.service.model.constraints.ValidDoa"
          maxItems: 50
        libraries:
          type: array
          items:
            $ref: '#/components/schemas/Library'
          maxItems: 1000
        dependentFeatures:
          type: array
          items:
            $ref: '#/components/schemas/Feature'
          maxItems: 1000
        independentFeatures:
          type: array
          items:
            $ref: '#/components/schemas/Feature'
          maxItems: 1000
        sharedWithOrganizations:
          type: array
          items:
            $ref: '#/components/schemas/Organization'
        visibility:
          $ref: '#/components/schemas/ModelVisibility'
        task:
          $ref: '#/components/schemas/ModelTask'
        torchConfig:
          type: object
          maxProperties: 20
          additionalProperties:
            $ref: '#/components/schemas/AnyValue'
        preprocessors:
          type: array
          items:
            $ref: '#/components/schemas/Transformer'
          maxItems: 50
        featurizers:
          type: array
          items:
            $ref: '#/components/schemas/Transformer'
          maxItems: 50
        rawModel:
          type: string
          format: byte
          description: A base64 representation of the raw model.
        creator:
          $ref: '#/components/schemas/User'
        canEdit:
          description: If the current user can edit the model
          type: boolean
        isAdmin:
          type: boolean
        selectedFeatures:
          type: array
          items:
            type: string
          maxItems: 1000
        tags:
          type: string
          maxLength: 1000
        legacyPredictionService:
          type: string
        scores:
          type: object
          properties:
            train:
              $ref: '#/components/schemas/Scores'
            test:
              $ref: '#/components/schemas/Scores'
            crossValidation:
              $ref: '#/components/schemas/Scores'
        extraConfig:
          $ref: '#/components/schemas/ModelExtraConfig'
        createdAt:
          type: string
          format: date-time
          description: The date and time when the feature was created.
          example: '2023-01-01T12:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: The date and time when the feature was last updated.
          example: '2023-01-01T12:00:00Z'
    ModelSummary:
      type: object
      required:
        - id
        - name
        - visibility
        - type
        - sharedWithOrganizations
        - createdAt
      properties:
        id:
          type: integer
          format: int64
          example: 0
        name:
          type: string
          minLength: 3
          maxLength: 255
          example: My Model
        visibility:
          $ref: '#/components/schemas/ModelVisibility'
        description:
          type: string
          minLength: 3
          maxLength: 50000
          example: A description of your model
        creator:
          $ref: '#/components/schemas/User'
        type:
          $ref: '#/components/schemas/ModelType'
        dependentFeaturesLength:
          type: integer
        independentFeaturesLength:
          type: integer
        sharedWithOrganizations:
          type: array
          items:
            $ref: '#/components/schemas/OrganizationSummary'
        createdAt:
          type: string
          format: date-time
          description: The date and time when the feature was created.
          example: '2023-01-01T12:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: The date and time when the feature was last updated.
          example: '2023-01-01T12:00:00Z'
    Scores:
      type: object
      properties:
        regression:
          $ref: '#/components/schemas/RegressionScores'
        binaryClassification:
          $ref: '#/components/schemas/BinaryClassificationScores'
        multiclassClassification:
          $ref: '#/components/schemas/MulticlassClassificationScores'
...
    ModelType:
      type: string
      enum:
        - SKLEARN
        - TORCH_ONNX
        - TORCHSCRIPT
        - R_BNLEARN_DISCRETE
        - R_CARET
        - R_GBM
        - R_NAIVE_BAYES
        - R_PBPK
        - R_RF
        - R_RPART
        - R_SVM
        - R_TREE_CLASS
        - R_TREE_REGR
        - QSAR_TOOLBOX_CALCULATOR
        - QSAR_TOOLBOX_QSAR_MODEL
        - QSAR_TOOLBOX_PROFILER
...
    Transformer:
      type: object
      description: A preprocessor for the model
      required:
        - name
        - config
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: StandardScaler
        config:
          type: object
          maxProperties: 20
          additionalProperties:
            '$ref': '#/components/schemas/AnyValue'
    ModelVisibility:
      type: string
      enum:
        - PUBLIC
        - ORG_SHARED
        - PRIVATE
    ModelTask:
      type: string
      enum:
        - REGRESSION
        - BINARY_CLASSIFICATION
        - MULTICLASS_CLASSIFICATION
    Library:
      type: object
      required:
        - name
        - version
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Library Name
        version:
          type: string
          example: 1.24.0
        createdAt:
          type: string
          format: date-time
          description: The date and time when the feature was created.
          example: '2023-01-01T12:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: The date and time when the feature was last updated.
          example: '2023-01-01T12:00:00Z'
    Doa:
      type: object
      required:
        - method
        - data
      properties:
        id:
          type: integer
          format: int64
        method:
          type: string
          enum:
            - LEVERAGE
            - BOUNDING_BOX
            - KERNEL_BASED
            - MEAN_VAR
            - MAHALANOBIS
            - CITY_BLOCK
          example: LEVERAGE
        data:
          type: object
          description: The doa calculated data
          oneOf:
            - '$ref': '#/components/schemas/LeverageDoa'
            - '$ref': '#/components/schemas/BoundingBoxDoa'
            - '$ref': '#/components/schemas/KernelBasedDoa'
            - '$ref': '#/components/schemas/MeanVarDoa'
            - '$ref': '#/components/schemas/MahalanobisDoa'
            - '$ref': '#/components/schemas/CityBlockDoa'
        createdAt:
          type: string
          format: date-time
          description: The date and time when the feature was created.
          example: '2023-01-01T12:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: The date and time when the feature was last updated.
          example: '2023-01-01T12:00:00Z'
...
    Dataset:
      type: object
      required:
        - type
        - input
        - entryType
      properties:
        id:
          type: integer
          format: int64
          example: 1
        type:
          $ref: '#/components/schemas/DatasetType'
        entryType:
          type: string
          enum:
            - ARRAY
          example: "ARRAY"
        input:
          type: array
          items: { }
          maxItems: 100
        result:
          type: array
          items: { }
        status:
          type: string
          enum:
            - CREATED
            - EXECUTING
            - FAILURE
            - SUCCESS
        failureReason:
          type: string
        userId:
          type: string
        modelId:
          type: integer
          format: int64
        modelName:
          type: string
        executedAt:
          type: string
          format: date-time
        executionFinishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DatasetCSV:
      type: object
      required:
        - type
        - inputFile
      properties:
        id:
          type: integer
          format: int64
          example: 1
        type:
          $ref: '#/components/schemas/DatasetType'
        inputFile:
          type: string
          format: byte
          description: A base64 representation in CSV format of the input values.
        values:
          type: array
          items: { }
        status:
          type: string
          enum:
            - CREATED
            - EXECUTING
            - FAILURE
            - SUCCESS
        failureReason:
          type: string
        modelId:
          type: integer
          format: int64
        modelName:
          type: string
        executedAt:
          type: string
          format: date-time
        executionFinishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AnyValue:
      description: Can be any value - string, number, boolean, array or object.